Show pending image uploads and allow cancling
Summary:
See Previous diffs for context about image uploading.

Now that we have a toolbar on the text field, we can add our upload progress to that toolbar, and even let you click on it to cancel the ongoing uploads.

We can accumulate this state from the ongoing uploads we know about. If any are uploading, we show a progress message.

If we click to cancel, we don't even truly cancel the uploads (because that might be fruitless), we just delete the placeholder texts and remove the in progress state.

Reviewed By: quark-zju

Differential Revision: D43332599

fbshipit-source-id: 8006ca752853167abac47aaea8a54a997c5b30f5