Separate CommitInfoField
Summary: Another small refactor, separate CommitInfoField into its own file. We will be expanding the functionality of this text field quite a bit, so separating it makes it easier to follow, and the main commit info file can just render it as a text field without worrying about the details.

Reviewed By: quark-zju

Differential Revision: D43332601

fbshipit-source-id: a0c155690f87e95a770c3473d65051372371a73f