Only enable image uploading internally
Summary:
Unfortunately, the `gh` cli does not currently support uploading images for PR descriptions. Doing so requires acting from the user's browser using their session cookie. It's technically possible for us to get users to use that cookie to make requests, but it's by no means an easy process.

This seems to be an often-requested feature of the gh cli, but they seem to not be interested in adding this feature:
https://github.com/cli/cli/issues/1895#issuecomment-718899617
for abuse and technical reasons.

Perhaps in the future we could support users configuring their own uploadable endpoints like a custom S3 bucket. Another variation would be to upload images to the repository via git push, then reference those file handles.

For now, we'll just disable image uploading features on non-internal builds.

Reviewed By: quark-zju

Differential Revision: D43369800

fbshipit-source-id: ef4cf083f6433bc795f2bf3488f476bed975b20c