Show error badge if image upload fails
Summary: If your image upload fails, we should let you know rather than just hiding the placeholder. We can do this with an inline error badge next to where the upload button is. Hovering shows the full error that was encountered. There's also a dismiss button to hide the error if you don't care about it.

Reviewed By: quark-zju

Differential Revision: D43411126

fbshipit-source-id: 1d6f029fd623c8dbcb54209a8e9dcb575c420bdd