Better mocks for VSCodeTextArea
Summary:
To support adding tests for image uploading, we need to refine our mocks for VSCodeTextArea.

In production, we use VSCodeTextArea from the vscode-webview-ui-toolkit library. These are implemented using the microsoft fast framework, which creates custom HTML components, like `vscode-text-area`, which means they have a shadow DOM component.

For TextArea, this means the DOM structure is like this:
```
<vscode-text-area>
  <label>...</label>
  <textarea part="control"></textarea>
</vscode-text-area>
```
(with some other properties I left out).

So a ref of VSCodeTextArea will actually point to the outer `<vscode-text-area>`. But almost always we want to do stuff to the underlying textarea. (e.g. typing, focus). Some things work on either, but it's kind of inconsistent.

<tangent>
TBH, I don't like custom components when using react, it's just added complexity for minimal styling isolation benefit.
</tanget>

In our tests, we can't actually use the real VSCodeTextArea, due to some incompatibility with how it's bundled and how jest loads our code. Ideally we'd be able to fix this, but it doesn't seem to work just yet.

So we have a mock of our VSCodeTextArea component (all of our VSCode webview ui toolkit components, actually).

This mock could go through the efforst to make a custom component.... but jest doesn't support them well regardless.

Previously, we were being janky by just returning a <textarea> directly. That works until it doesn't. We started hitting situations where something would work in production but not in tests, because our tests didn't have the extra wrapper of the custom component.

Instead, we do something to mimic the extra structure. So in this diff I restruct our mock to wrap a text area:
```
<div>
  <textarea part="control"></textarea>
</div>
```
This more closely matches the real structure of `<vscode-text-area>`. This means our tests can access the inner textarea while the rest of our code paths treat it as a real VSCodeTextArea wrapper.

In this diff I also clean up some references to `HTMLInputElements`, which are all actually `HTMLTextAreaElement`s.

Reviewed By: quark-zju

Differential Revision: D43408511

fbshipit-source-id: 1bd5c4638dca8cb87ce5ff6d145dbd12bcd103e9