Support opening diff view in VS Code from ISL
Summary:
Add a platform hook to open files in a platform-native diff view.

If this platform API is implemented, then we'll show a diff view icon next to each uncommitted change.

We currently only show this on the uncommitted changes section (not changes on each commit you select)

This API requires passing a specific comparison. Even though we only use uncommitted changes right now, this will let us use this feature elsewhere in the future.

Reviewed By: quark-zju

Differential Revision: D43468032

fbshipit-source-id: 8406f2821c2e304b9066a8a3de83941f1e065557