Add button in SCM sidebar to open diff view
Summary:
This was requested in https://github.com/facebook/sapling/issues/531.

We should let you click to open the diff view on changed files in the vscode SCM sidebar. We already support clicking to open a file. The diff view command is already implemented, so we just need to call it and handle SourceControlResourceStates.

Reviewed By: quark-zju

Differential Revision: D43477399

fbshipit-source-id: a3148ce1367778206139b3c8fdfe69a412fa5fbd