add Python stub for `prefetch-profile`
Summary:
In order for the command to show up in our help text, we must add a stub and print out "not-implemented" if someone somehow manages to invoke the Python version of `prefetch-profile`.

Instead of adding stubs for each sub-subcommand for prefetch-profile, we'll simply redirect the user to the Rust help text.

Reviewed By: fanzeyi

Differential Revision: D43482266

fbshipit-source-id: f29d511bfb06a6d1cc74386cc88e685024384ece