inodes: only unload when inode isn't being loaded
Summary:
A very long standing bug in EdenFS is caused by inodes completing to load but
that inodes not being found in the unloadedInodes_ map in the InodeMap. While
the root cause has never been identified, reading the code shows that there are
very few places removing inodes from the unloadedInodes_ map. One of them
appears to be when the FS refcount is decremented, if somehow that inode is
being loaded, we can end up in a situation where the inode will not be found in
the InodeMap, causing EdenFS to crash.

Reviewed By: MichaelCuevas

Differential Revision: D43484128

fbshipit-source-id: 6208ba74fbf02f21dd67144abf0302db10ae318a