Fix jitter-delay for SQL Blob
Summary: When adding jitter-delay for SQL Blob I/O ops, we select a random value based on a range between 0 secs and `lag` secs. However, the raw `lag` seconds provided as input can be 0 in which case the range becomes empty and the random delay generator panics (Example: P628110845). This diff ensures that the upperbound for the delay doesn't go beyond 1 second and the lower bound stays above 0 seconds.

Reviewed By: mitrandir77

Differential Revision: D43435451

fbshipit-source-id: 4ec6f39fde6b1541abbcef24f8fba1c9fb9245ad