NextSteps stream
Summary:
This adds implementations of `next_step` and `try_next_step`, which are similar to `then` and `and_then` in `(Try)StreamExt`, but they continue awaiting for the next element while the processing function processes the previous one.

This is pretty useful for optimising assembly line like stuff, like most of our things. E.g. in mirror_hg_commits you don't need to wait for the previous commit to write the mappings before starting to copy the next commit to the blobstore.

It **does not** parallelise anything in the same level. Both "levels" still happen in order. I could make the "queue" be bigger than 1, but I don't know how useful that is.

Reviewed By: mitrandir77

Differential Revision: D43395160

fbshipit-source-id: 050a90a4f3a0122cf2d6df6ab755f7691c4b6a3a