Rename BookmarkName to BookmarkKey
Summary:
In a future diff, we want to separate the concepts of `BookmarkName` which represents the name of a bookmark with that of `BookmarkKey`, which refers to a unique bookmark, namespaced by a category: `Branch`, `Tag` or `Note`.

This will allow us to represent different types of refs importing from git repositories that may have conflicting names, but are semantically different.

Since `BookmarkName` is widely referred to in mononoke, start with a pure renaming diff for ease of reviewing.

This code change is a pure refactoring and has no effect on the behaviour of Mononoke.

Reviewed By: RajivTS

Differential Revision: D43350002

fbshipit-source-id: e64a84f7b3b70c4254702ebd7ace4d645cd00eff