Distinguish between BookmarkKey and BookmarkName
Summary:
In the parent diff, we renamed `BookmarkName` to `BookmarkKey`, reflecting the fact that most places that previously used `BookmarkName` actually seek a unique way to refer to a bookmark, which is represented by a `BookmarkKey` in the `Git` compatible view of the world.

We now introduce again the concept of `BookmarkName`, which represents the name of a bookmark (wrapper around an ascii string) and does not uniquely identify it.

With this semantic difference in place, we refer to `BookmarkKey` or `BookmarkName` as appropriate in the codebase.

A future diff will introduce the concept of `BookmarkCategory` to meaningfully differentiate between a `BookmarkName` and a `BookmarkKey`. For now, we are only using types as a way to express the semantic difference between these two concepts. This separation of diffs is somewhat arbitrary and is meant to facilitate review of each diff.

This code change is a pure refactoring and has no effect on the behaviour of Mononoke (barring some stdout changes reflecting in integration tests).

Reviewed By: RajivTS

Differential Revision: D43350001

fbshipit-source-id: 6f314cd6838cbc4ffd59c41e874d1fa997a627a3