Revert button in SCM sidebar
Summary:
Add a revert button to the SCM sidebar. This gives you an extra bit of power to this sidebar.

This is the first command we're adding that runs an Operation. We are better set up to run operations from the ISL webview. We have access here, we just need to pipe the data through in the same way the webview does. We don't currently send the progress data back to any open webviews. This means you won't see the progress. But we could probably support this in the future.

Reviewed By: quark-zju

Differential Revision: D43484745

fbshipit-source-id: 30978d843068223b43d846fb13225ed1669d2697