Typed command running and tracking
Summary:
Rework our VS Code command registration & execution. Define commands in a top-level object, so we statically know the types of our registered commands (parameters and return type). This lets us have type-safe programmatic execution of commands.

With a little care, we can even pass a Context object to our command handlers. This lets us pass global state that applies to all commands. The salient example is the tracker instance for this VS Code session, for analytics (internal only, this is a no-op in OSS builds of course).

We can also just generally use the tracker to track commands being run, so we get an idea of how people use the vscode extension internally.

Reviewed By: quark-zju

Differential Revision: D43484746

fbshipit-source-id: 5ced7bcf74e656626fb6435e9da59ace701b4904