Create globalrevs when backsyncing
Summary:
This adds handling of backsyncing to small repos with globalrevs. It will be necessary for www merge, as it uses globalrevs.

On the large repo, the globalrevs will not exist, but they will be created when backsyncing to the small repo.

Let's make the changes clear. Globalrevs are a different identifier for commits, which are sequential numbers, which only work on 1 bookmark and for repos without merges. They could be but are not derived data, and globalrev information is stored INSIDE the bonsai changeset itself, so we can't derive it after the fact. Globalrev is also stored in a DB that allows mapping bonsai <> globalrev <> hg

This is how globalrevs currently work:
- We have a pushrebase hook, that during pushrebase on master in www, reads the latest globalrev number from the DB, then modifies the changesets being pushrebased and adds the globalrev info on them.
- That hook also stores the globalrev mappings on the DB.

This is how they'll work on the megarepo (with this diff):
- When landing/pushrebasing to fbsource, we don't do anything related to globalrevs.
- When backsyncing a commit from fbsource to www, we store globalrev info on every commit, as we can't know for sure if it's eventually going to become part of master. We don't store anything in the mapping DB.
- When moving bookmarks as part of the backsync, we store the information from the new commits in the globalrev DB.

This should work mostly the same way in practice, with these exceptions:
- Landing to other bookmarks (not really a thing on www), or even draft commits that eventually get backsynced will have globalrev info on them, though not on the DB.
- If we roll back the master bookmark and then move it forward again, we will REPLACE the globalrevs. Currently the behaviour is to have gaps.
  - In the figure below, if master was C and ABC had globalrevs 1 2 3, but then master was changed to B and then D, then D would have globalrev 4 in the old behaviour and 3 in the new behaviour.

```
C D
| /
B
|
A
```

Reviewed By: mitrandir77

Differential Revision: D43004874

fbshipit-source-id: 552e2ad26eafe5abefe79511bc31a3f052b84052