Add test for globalrevs in megarepo
Summary:
This adds a test that uses megarepo stuff where the small repo uses globalrevs. It does the forward sync phase, then the backward sync one, and checks the globalrevs during the process.

It verifies that everything continues working with forward sync, but stops when the source of truth is changed. This is what I plan to fix on the next diffs, and this test will show it.

Differential Revision: D43159308

fbshipit-source-id: 714adf3e9ac63c61233512b8f9846adcd53ccaec