filenodes: remove FilenodeRangeResult
Summary:
The `FilenodeRangeResult` type conflates two concepts: a range being too large, and the result of filenodes lookups being unavailable because filenodes are disabled.

Replace it with `FilenodeResult<FilenodeRange>`, keeping the notion of whether or not filenodes are disabled to `FilenodeResult`, and handling this concept always through that.

The type was also pointlessly generic.  It almost always contained `Vec<FilenodeInfo>`, except in a couple of places where that is easily removed, so remove the generic parameter.

Differential Revision: D43410852

fbshipit-source-id: 6be34555b8b56b72d2ea4b071fb03c64648e8406