service: make EdenServer::getMountPoints safer
Summary:
In a few cases, the code wrongly assumes that the mount points returned by
getMountPoints will be safe to access even though it is documented (and
possible) that the mount will be shutdown after getMountPoints returns. To
prevent this, collecting the root inode with the mount point lock held is
necessary, thus the function now does exactly this.

Reviewed By: kmancini

Differential Revision: D43516295

fbshipit-source-id: a4bc48a60c187f1610e05803b06abdd8793533b3