Trim full path to binary from minitop
Summary:
On Windows, minitop should hide the absolute path to processes' binaries to make the process filename more visible.

There's a shlex crate that can parse Unix-like command lines, but I didn't see an equivalent for Windows. (The Unix command line grammar is different, and in particular shlex will eat unquoted backslashes that are legal on the Windows command line.) So here we use the `CommandLineToArgvW` win32 function to get the Windows equivalent.

After parsing the command line and stripping the full path from the executable name, for now we just naively join the arguments back together, which means binary names containing spaces won't be properly quoted. I may add this in a following diff.

Reviewed By: xavierd

Differential Revision: D43509750

fbshipit-source-id: d56535ded4984a8106f4a26dd5d1b54e2c60e297