Back out "add Python stub for `prefetch-profile`" and "[edenfs] kill python prefetch-profiles"
Summary: Some environments are still somehow using the Python version of edenfsctl even though they should be fully on the Rust version. That means some environments are experiencing failures due to `eden prefetch-profile` not being a valid command. This diff makes the Python version of edenfsctl available again to help fix this.

Reviewed By: fanzeyi

Differential Revision: D43589012

fbshipit-source-id: eb529b48bcadd734f80e809d2f7b41ef081d7442