Allow open_managed_repos to open default repo-set in non-sharded mode
Summary: `open_managed_repos` method in `MononokeApp` opens the set of filtered and enabled repos for the given service. It also considers the deep-sharding status of the repos while opening them during initialization. If all the repos for the service are deep-sharded, then none of them will be opened during startup. This is desired when the service is running in sharded mode on TW, but not so when it is being run locally for testing. This diff adds a check to ensure that deep-sharding status is only considered if the service is executing in sharding mode.

Differential Revision: D43618886

fbshipit-source-id: 23ac5bcea7451f5e81ecb2dd74d61860bb3ef50f