workingcopy: reduce usage of Arc<Mutex<TreeState>>
Summary:
There isn't any parallelism, so this type doesn't make sense conceptually. It is used for convenience so various helper structs can hold a mutable TreeState without dealing with references. However, it leads to code such as the FileChangeDetector which needs to lock the mutex O(changed files) times.

Instead, let's just pass `&mut TreeState` as a parameter. The TreeState a fundamental input to most code in here, so conceptually this makes sense to me.

Reviewed By: quark-zju, zzl0

Differential Revision: D43501748

fbshipit-source-id: 3320b0dade2ffc3949e65b91b7adcb3368366d7b