caching: add cache mode parameter to command line with local-only option
Summary:
Add a new `--cache-mode` parameter to all mononoke command line programs.  This replaces the existing `--skip-caching` option (which is now `--cache-mode=disabled`) and `--cachelib-only-blobstore` (which is now deprecated).

A new `--cache-mode=local-only` mode is added, which enables cachelib for all caches (not just the blobstore), but keeps memcache disabled.  This should be used in preference to `--cachelib-only-blobstore` when shared caching is not desired, as it will still enable local cachelib-based caching to work.

Reviewed By: RajivTS

Differential Revision: D43444275

fbshipit-source-id: d46b4dd7a7de3e7d7dcdf1f40899d49bb309895b