Refactor test if operation is running
Summary:
Small refactor preparing for the next diff.

We will likely want to support checking if a particular operation is running or queued in different places in the UI. For example, while `pull` is running, we shouldn't let you run or queue pull again (just to make it clear that the UI got your request to run it).

So I'm extracting a hook that makes this check pretty easy to do in different places. You provide the Operation you care about in particular and it returns if its queued or running.

I use a union of `'queued' | 'running' | undefined` in the return type (as opposed to `'queued' | 'running'  | 'not_running'`) so you can easily use a falsey test for `=== 'queued' || === 'running'`.

Reviewed By: quark-zju

Differential Revision: D43670011

fbshipit-source-id: b5ecca7cc224cb0e85c0c525d0e404685e146919