Disable continue/abort merge buttons while running those operations
Summary:
In D43670012 we added support for optimistic state when running resolve commands. Additionally, we want to make the continue/abort buttons behave more responsively than they do right now, while not using optimistic state to dismiss the buttons entirely (see D43670012 for reasoning).

The idea is that we disable the continue/abort merge buttons while continue/abort are running or queued to run. This way the button immediately takes effect in the UI. When they finish, there's still a moment where the command has run but we haven't gotten new conflict state back (the age old problem which is why we have the whole optimistic state system). In this case, we don't have optimistic appliers to check when to consider the operation completed and the optimistic state resolved, so we should just keep disabling the buttons while the most recent command is a successful continue/abort.

This makes the buttons remain disabled until the conflicts disappear as expected.

This should be safe:
- hitting additional conflicts will make `sl continue` return non-zero, so the button will not be disabled to try again
- successful abort should always make conflicts disappear
- successful continue should also make conflicts disappear, further commands would displace the current operation so this would no longer take effect

Reviewed By: quark-zju

Differential Revision: D43670010

fbshipit-source-id: 2c55de634f07e903227c3e60b6bcd8d9a5332653