Add optimistic state for merge conflicts
Summary:
We support optimistic state for changes to the commit graph and uncommitted changes. We want to extend this support to merge conflicts.

This will allow us to have a very smooth UX when handling merge conflicts. Namely, clicking the resolve button should immediately act as though the file is resolved and if all files are resolved it should let you run `sl continue`.

This diff adds support for this. We just add a new type of optimistic state that Operations may define. It works very similarly to Uncommitted changes optimistic state, but defines the entire MergeConflicts data structure (mostly we care about the list of files).

We use this optimistic applier function for the ResolveOperation, used when clicking the resolve checkmark next to files.

We don't use it for continue/abort, although it would techincally be possible. The issue is that we may hit additional conflicts or other weird situations, and I could imagine this causing unreliability of the continue/abort buttons. We want those to work very transparently, so I opt to not make them optimistically hide conflicts.

Reviewed By: quark-zju

Differential Revision: D43670012

fbshipit-source-id: ba3e36a6fd75993bfccc53634cf9a5d7f16814c2