assert our fsck assumptions about directories
Summary:
Currently, fsck will create an inode if a file is missing from disk, but present
in the source control tree. This matches the behavior of what PrjFS will
automatically do for files with in placeholder directories. But it doesn't match
for full directories.

To ensure the inode state matches the on disk state after fsck we need to detect
placeholder vs. full directories.

Fortunatly, the recall attribute seems to do just that.

This diff asserts the assumptions we are making about directorys. So that we
can safely use the recall attribute to distinguish placeholder and full
directories above.

In the above diffs, I will teach fsck to handle remove files in full add
placeholder directories the differently, so that the inide state matches disk.
Ultimatly I would like it all to behave the same: removed files stay removed.
But preventing PrjFS from bringing back files in placeholder directories is
difficult. So we are starting with correcting eden's behavior for full
directories. Currently, fsck will create an inode if a file is missing from disk, but present
in the source control tree. This matches the behavior of what PrjFS will
automatically do for files with in placeholder directories. But it doesn't match
for full directories.

Here I am teaching fsck to handle remove files in full and placeholder
directories differently. This change ensures that EdenFS's internal state always
matches that of disk after restart.

Notably this does not prevent files from reappearing after they are deleted.
If the parent directory is a placeholder and a file is deleted on disk while
eden is stopped, that file will come back after eden starts.

Ultimatly I would like all removals to behave the same: removed files stay
removed.

But preventing PrjFS from bringing back files in placeholder directories is
difficult. So we are starting with correcting eden's behavior for full
directories.

Reviewed By: xavierd

Differential Revision: D43379229

fbshipit-source-id: 0920834980f6ea6e76214555c196f355a847b8b1