ui: don't default username to <user>@<fqdn>
Summary: If the user hasn't configured their ui.username, I would rather the command error out instead of using <user>@<fqdn>. On Windows in particular, this default gives a very wrong user name that gets immortalized in the commit cloud workspace. It's much less toil for the user to configure the username early.

Reviewed By: quark-zju

Differential Revision: D43668690

fbshipit-source-id: 64c4e0d235b2fe2fa80c76da6e04917573989a9b