Create BookmarkCategory
Summary:
# TL;DR

Add a `BookmarkCategory` enum for the purpose of namespacing bookmarks with the same `BookmarkName`. Propagate the change from the `bookmarks` table in the database to all of Mononoke.

In this change, only `BookmarkCategory::Branch` is used when creating new bookmarks, which makes it essentially a no-op, but allows us to extend the functionality as we propagate the change through interfaces such as `SCS`.

# Some context

As we start importing repositories from Git, we need to make sure that the Mononoke data model is able to represent anything we import.

In git, refs live in different "namespaces" - subdirectories in the `.git` directory (e.g: `refs/heads`, `refs/tags`, `refs/remotes`, ...)
For an exhaustive list of categories of refs, documented by our friend Byron from gitoxide, see [the git-ref crate's documentation](https://docs.rs/git-ref/0.24.1/git_ref/enum.Category.html).

Of all of these kinds of refs, we only actively care about the kind of refs that we might need to import.
This boils down to 3 categories: `Branch`, `Tag`, `Note`.

* `Branch`
Used for a git branch (imported from a ref under `refs/heads` or `refs/remotes`) or a mononoke bookmark. It's the "default" behaviour for a bookmark. It points to a `BonsaiChangeset` that represents a git commit if imported from Git, or a `BonsaiChangeset` representing a mercurial changeset.
* `Tag`
Used for a git tag (imported from a ref under `refs/tags`). It points to a `BonsaiChangeset` that represents a commit (lightweight tag) or a  `BonsaiChangeset` that represents an annotation, and that points to a `BonsaiChangeset` representing a commit (annotated tag).
* `Note`
Used for a git note. They're a special kind of annotated tag that live under `refs/notes`. They can be created with `git notes` and `git` uses the annotations to display it at the end of the commit message. This is a mechanism for appending to a commit message without rewriting the history.

`Pseudoref`s (like `HEAD`) and `worktrees` are client features of git and are not supported in Mononoke.

Reviewed By: RajivTS

Differential Revision: D43444397

fbshipit-source-id: 0b0b5f18b30ae1f9871d2402cb2fc21fe3ebd892