Introduce Backfill mode in AliasVerify tool
Summary: I plan to use `AliasVerify` tool to backfill `ContentMetadataV2` and generate the new `SeededBlake3` aliases for the files. This diff adds the `Backfill` mode along with the relevant mutable counters which would be used to determine where to start the backfilling from, in case the task restarts in between.

Reviewed By: yancouto

Differential Revision: D42884327

fbshipit-source-id: bb610e5cf8f30848bc74f568f3151132ca35e7f6