Add Seeded Blake3 in ContentMetadataV2
Summary: This diff introduces `seeded_blake3` hash as a new property in ContentMetadataV2 and modifies integration tests accordingly. It will be stored when a new file is created or when the metadata for a file is missing. Currently, the client cannot request this specific hash and this hash can't serve as an alias for file content stored in Mononoke. Those changes will be made in follow-up diffs.

Differential Revision: D42801488

fbshipit-source-id: 45ce9ac1b7a7c3726e0ba18dc99dfa621fee3da2