Add support for ContentAlias in Blobstore Fetch
Summary: As in title. Blobstore fetch is a tool used to output formatted blobs based on the specified key. In the list of supported types, we did not have ContentAlias. When backfilling `ContentMetadataV2` and associated `SeededBlake3` alias, I found it useful to rely on the blobstore fetch tool to validate if the aliases have been created properly. This diff adds support for `ContentAlias` in the tool so that future users can check if the alias has been properly created.

Differential Revision: D42969200

fbshipit-source-id: 7d3802561cef904f721014bbd90af3d602085a64