Modify AliasVerify to use reloadable repos
Summary: Instead of creating and owning a static copy of the repo, the command now fetches the repo from `MononokeRepos` which ensures that it will always get the latest state of the repo and associated configs. This will also aid in sharding `AliasVerify` which would be done in the next diff.

Differential Revision: D42996407

fbshipit-source-id: ca0b65cb861bafbdd346e36f0ff10b7414080c18