Implement Blake3 as an Alias for fetching files
Summary: Now that the `seeded_blake3` hash is part of `ContentMetadataV2`, we can use it as an alias to fetch the file corresponding to this hash. This hash can now also be used as a means to validate correct upload or store operation (i.e. hash of sent file == hash of uploaded file).

Reviewed By: yancouto

Differential Revision: D42804812

fbshipit-source-id: ab9c196c9e11404881d278a5014df5b3972c8fcb