Add support for Blake3, Sha1 and GitSha1 aliases in AliasVerify tool
Summary: The `alias_verify` command works in two modes, `verify` and `generate`. In `verify` mode, it checks if the Sha256 hash of the file exists as an alias for it and reports an error if it doesn't (or if an alias exists but with a different hash). In `generate` mode, if the alias is found missing, it generates the alias for that file. Currently, the command works with only `Sha256` hashes, but a file alias can be `Sha1`, `GitSha1` and `SeededBlake3` in addition of `Sha256` hash. This diff adds support for the remaining types of aliases so that the tool can be used to verify and generate any possible alias.

Reviewed By: yancouto

Differential Revision: D42844271

fbshipit-source-id: 11ebd7356f85037915d4137cd8a38d8e41d24c94