Expose linelog API to edit previous revisions
Summary:
Extend the `recordText` API to support editing previous revisions by providing
the desired content at the desired rev. This can potentially be used in a live
editor in stack editing.

The benefit of using linelog is that:
- If the change stays unchanged by newer revisions, then the change applies to
  newer revisions.
- If the change is deleted (or changed) by newer revisions, then it remains
  deleted in the newer revisions.

Note: certain edits (especially deleting, or replacing a large chunk of text)
can be ambiguous and confusing. They are "legit" conflicts in some sense but
we currently don't surface the conflicts or disallow certain kinds of edits.

Reviewed By: muirdm

Differential Revision: D43773254

fbshipit-source-id: d03237abe9d443ee9f6ac46b0afdf69234e3f3eb