Add specific error message when no commits found
Summary:
Previously, if `sl log` returned no commits, we'd send `[]` back to the client, which would just keep thinking we hadn't yet initialized the set of commits and it would show a spinner forever.

instead, let's send an error back to the client if we don't get any commits.

This could happen in two cases I know of:
1. the repo is empty (no initial commit exists)
2. `sl log --rev smartlog()` returns nothing, because it's not tracking any remote bookmarks and there's no commits within the last 2 weeks otherwise.

for (1), we can add a button that creates an empty Initial commit for you.

For (2), need to ensure we always fetch the head commit so it always tries to load at least one commit

Reviewed By: quark-zju

Differential Revision: D43785053

fbshipit-source-id: 1c1b1c4f5d1db6a78341e27b6ec0c0b92a0d6e53