Fix pull button rendering relative date with -Infinity
Summary:
Bug report came in saying ISL would spin forever and sometimes crash.

de-minified stack trace:
```
Something went wrongThis should never be reached.Error: This should
    at relativeDate (../../isl/src/PullButton.tsx:33:4)
    at c (../../node_modules/react-dom/cjs/react-dom.production.min.js:167:136)
    at Xh (../../node_modules/react-dom/cjs/react-dom.production.min.js:290:336)
    at Wk (../../node_modules/react-dom/cjs/react-dom.production.min.js:280:388)
    at Vk (../../node_modules/react-dom/cjs/react-dom.production.min.js:280:319)
    at Uk (../../node_modules/react-dom/cjs/react-dom.production.min.js:280:179)
    at Jk (../../node_modules/react-dom/cjs/react-dom.production.min.js:271:87)
    at Ok (../../node_modules/react-dom/cjs/react-dom.production.min.js:273:299)
    at d (../../node_modules/react-dom/cjs/react-dom.production.min.js:127:104)
```

relativeDate was crashing because we assumed no date input would not be `< Number.MAX_VALUE`. However, if we got no commits back from `sl`, we'd take `Math.max(...[])`, which is `-Infinity`. Then, in computing the date delta, we'd get `+Infinity`, which is not `< Number.MAX_VALUE`.

To guard against this, we shouldn't try to render the relativeDat if there are no commits. We could also maybe change relative date to not throw if you give it -Infinity. But also... we should just shouldn't do that.

Reviewed By: quark-zju

Differential Revision: D43784765

fbshipit-source-id: 03637738f132281da90184f047e23be1423dfd7a