Always fetch head commit
Summary: If you open a repo you haven't looked at in a few weeks, your head commit will be > 14 days old, and our computed revset to fetch from hg log will be `smartlog() and date(-14)`, which means even head won't be fetched. We need to always include the head commit in this, so that we fetch some commit. This should practically prevent us from returning `[]` on a repo that has some commits in it.

Reviewed By: zzl0

Differential Revision: D43786238

fbshipit-source-id: 7c4a64f92c7022acaf619eb7032a16ace93ef3cb