make error message with gpg more informative
Summary:
As an [user pointed out in #544](https://github.com/facebook/sapling/issues/544), `sl` crashes when `gpg` returns an error code instead of returning a slightly more useful error message.

This diff also changes the gpg flags so that it doesn't error out due to lack of user input. Previously it would return an error 2 when it waited for user input.

Reviewed By: zzl0

Differential Revision: D43860003

fbshipit-source-id: ecfcb6f8cabb736365ea407211437c52f878f0c8