Add scuba logging when ContentMetadataV2 gets repaired
Summary: Since the backfilling for ContentMetadataV2 and Blake3 happened in multiple phases, there can be deserialization error when reading an earlier version of stored metadata based on the current thrift structure. The code has already been modified to do the repair in such a case and this diff adds the relevant scuba logging so that all such repairs can be tracked.

Reviewed By: markbt

Differential Revision: D43870873

fbshipit-source-id: b0f83369152465f7afce354d096457e82c25a586