Enable chunk-based backfilling in AliasVerify
Summary: I plan on using the `AliasVerify` tool to perform backfilling of `ContentMetadataV2` and `SeededBlake3` hash. Since we need to cover all the commits and not just CommitCloud commits, the total number of commits to process is very very high (~850M). This diff adds chunking support in the tool so when executing in sharded setting, the sharded executor will provide `configerator_TO_1-32` as the input which would mean that the task should work on the first of the thirty-two chunks of configerator changesets. This would allow for parallelization and performance speed-ups.

Reviewed By: markbt

Differential Revision: D43118215

fbshipit-source-id: 2e0487684b3812e004b768e8c119d542662066b6