histedit: avoid pointless merge for "fold"/"roll"
Summary: The fold action collapsed commits up to and including the working copy, but it left the working copy dirty. When it subsequently updated to the new collapsed commit, it "merged" all the files. They merge cleanly since their contents are identical, but we still invoke the merge drivers (why are we invoking merge drivers if the contents are identical?). Fix by using ctx.markcommitted() which updates the dirstate based on what was committed. And, we no longer need to do the "hg.update()" anymore since the working copy will already be on the collapsed commit.

Reviewed By: quark-zju, zzl0

Differential Revision: D43718659

fbshipit-source-id: 26865cca11feb600e21933c60ed83796212da954