enable running integration tests on apple silicon
Summary:
We are unable to run any integration tests on apple silicon because we fail to
start the eden daemon.

We are expecting to be able to find arch, but with execve we need to pass the
full path. according to the python docs adding a "p" should allow the command to
resolve the path with the current environment.

This allows the tests to get farther on my M1.

Reviewed By: mshroyer

Differential Revision: D43923320

fbshipit-source-id: eb737eb561659bd8c9d3cd6bc214f8860b657294