Set TMPDIR to /tmp on macOS to enable test_status_socket
Summary:
On macOS, Python's `tempfile.mkdtemp` defaults to a long path under
`/private/var/folders`.  `test_status_socket` creates a Unix-domain socket within
an Eden mount, but because of the long path this currently fails on macOS.

There are a couple different places where `mkdtemp` is invoked in our
integration tests, so here we override the default path on macOS to `/tmp`, by
way of the `TMPPATH` environment variable.

Reviewed By: genevievehelsel

Differential Revision: D43925614

fbshipit-source-id: fdd59668fcfe68ba57fabd0a18ec14bb1b3079d6